Move single screens to features directory Rename profile presentation directory
