+++ lib/main.dart
@@ -3,15 +3,15 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
+import 'features/auth/presentation/view/auth_screen.dart';
 import 'features/sign_in/presentations/view/sign_in_screen.dart';
+import 'features/splash/presentation/view/splash_screen.dart';
 import 'firebase_options.dart';
 import 'repositories/repositories.dart';
 import 'services/auth_service.dart';
-import 'ui/screens/auth/auth_screen.dart';
 import 'ui/screens/home/home_screen.dart';
 import 'ui/screens/search/search_screen.dart';
 import 'ui/screens/sign_up/sign_up_screen.dart';
-import 'ui/screens/splash/splash_screen.dart';
 import 'ui/shared/theme/app_theme.dart';
 
 final authRepoProvider =
